Main category view taken care of.

+++ app/src/main/java/coljamkop/momoney/Content/Budgetable.java
@@ -1,9 +1,11 @@
 package coljamkop.momoney.Content;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Aghbac on 6/10/16.
  */
 public interface Budgetable {
     boolean isInGoal();
-    double getTotal();
+    BigDecimal getTotal();
 }
